baru

+++ psi_inventaris/routes/web.php
@@ -5,6 +5,8 @@ use SebastianBergmann\CodeCoverage\Report\Html\Dashboard;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\AdministrasiController;
 use App\Http\Controllers\BarangController;
+use App\Http\Controllers\LoginController;
+use App\Http\Controllers\TransaksiKeluarController;
 use App\Http\Controllers\TransaksiMasukController;
 
 /*
@@ -18,41 +20,54 @@ use App\Http\Controllers\TransaksiMasukController;
 |
 */
 
-Route::get('/', function () {
-    return view('login');
-});
+Route::get('/', [LoginController::class, 'index'])->name('login')->middleware('guest');
 
-Route::get('/dashboard', [DashboardController::class, 'index']);
+Route::post('/login', [LoginController::class, 'authenticate']);
+
+Route::post('/logout', [LoginController::class, 'logout'])->middleware('auth');
+
+Route::get('/dashboard', [DashboardController::class, 'index'])->middleware('auth');
 
 Route::group(['prefix'=>'administrasi'], function(){
-    Route::get('/', [AdministrasiController::class, 'index']);
-    Route::post('/list', [AdministrasiController::class, 'list']);
-    Route::get('/create', [AdministrasiController::class, 'create']);
-    Route::post('/', [AdministrasiController::class, 'store']);
-    Route::get('/{id}', [AdministrasiController::class, 'show']);
-    Route::get('/{id}/edit', [AdministrasiController::class, 'edit']);
-    Route::put('/{id}', [AdministrasiController::class, 'update']);
-    Route::delete('/{id}', [AdministrasiController::class, 'destroy']);
+    Route::get('/', [AdministrasiController::class, 'index'])->middleware('auth');
+    Route::post('/list', [AdministrasiController::class, 'list'])->middleware('auth');
+    Route::get('/create', [AdministrasiController::class, 'create'])->middleware('auth');
+    Route::post('/', [AdministrasiController::class, 'store'])->middleware('auth');
+    Route::get('/{id}', [AdministrasiController::class, 'show'])->middleware('auth');
+    Route::get('/{id}/edit', [AdministrasiController::class, 'edit'])->middleware('auth');
+    Route::put('/{id}', [AdministrasiController::class, 'update'])->middleware('auth');
+    Route::delete('/{id}', [AdministrasiController::class, 'destroy'])->middleware('auth');
 });
 
 Route::group(['prefix'=>'barang'], function(){
-    Route::get('/', [BarangController::class, 'index']);
-    Route::post('/list', [BarangController::class, 'list']);
-    Route::get('/create', [BarangController::class, 'create']);
-    Route::post('/', [BarangController::class, 'store']);
-    Route::get('/{id}', [BarangController::class, 'show']);
-    Route::get('/{id}/edit', [BarangController::class, 'edit']);
-    Route::put('/{id}', [BarangController::class, 'update']);
-    Route::delete('/{id}', [BarangController::class, 'destroy']);
+    Route::get('/', [BarangController::class, 'index'])->middleware('auth');
+    Route::post('/list', [BarangController::class, 'list'])->middleware('auth');
+    Route::get('/create', [BarangController::class, 'create'])->middleware('auth');
+    Route::post('/', [BarangController::class, 'store'])->middleware('auth');
+    Route::get('/{id}', [BarangController::class, 'show'])->middleware('auth');
+    Route::get('/{id}/edit', [BarangController::class, 'edit'])->middleware('auth');
+    Route::put('/{id}', [BarangController::class, 'update'])->middleware('auth');
+    Route::delete('/{id}', [BarangController::class, 'destroy'])->middleware('auth');
 });
 
 Route::group(['prefix'=>'transaksiMasuk'], function(){
-    Route::get('/', [TransaksiMasukController::class, 'index']);
-    Route::post('/list', [TransaksiMasukController::class, 'list']);
-    Route::get('/create', [TransaksiMasukController::class, 'create']);
-    Route::post('/', [TransaksiMasukController::class, 'store']);
-    Route::get('/{id}', [TransaksiMasukController::class, 'show']);
-    Route::get('/{id}/edit', [TransaksiMasukController::class, 'edit']);
-    Route::put('/{id}', [TransaksiMasukController::class, 'update']);
-    Route::delete('/{id}', [TransaksiMasukController::class, 'destroy']);
+    Route::get('/', [TransaksiMasukController::class, 'index'])->middleware('auth');
+    Route::post('/list', [TransaksiMasukController::class, 'list'])->middleware('auth');
+    Route::get('/create', [TransaksiMasukController::class, 'create'])->middleware('auth');
+    Route::post('/', [TransaksiMasukController::class, 'store'])->middleware('auth');
+    Route::get('/{id}', [TransaksiMasukController::class, 'show'])->middleware('auth');
+    Route::get('/{id}/edit', [TransaksiMasukController::class, 'edit'])->middleware('auth');
+    Route::put('/{id}', [TransaksiMasukController::class, 'update'])->middleware('auth');
+    Route::delete('/{id}', [TransaksiMasukController::class, 'destroy'])->middleware('auth');
+});
+
+Route::group(['prefix'=>'transaksiKeluar'], function(){
+    Route::get('/', [TransaksiKeluarController::class, 'index'])->middleware('auth');
+    Route::post('/list', [TransaksiKeluarController::class, 'list'])->middleware('auth');
+    Route::get('/create', [TransaksiKeluarController::class, 'create'])->middleware('auth');
+    Route::post('/', [TransaksiKeluarController::class, 'store'])->middleware('auth');
+    Route::get('/{id}', [TransaksiKeluarController::class, 'show'])->middleware('auth');
+    Route::get('/{id}/edit', [TransaksiKeluarController::class, 'edit'])->middleware('auth');
+    Route::put('/{id}', [TransaksiKeluarController::class, 'update'])->middleware('auth');
+    Route::delete('/{id}', [TransaksiKeluarController::class, 'destroy'])->middleware('auth');
 });
